requirements.txt
streamlit
pytesseract
Pillow
gtts
requests
langchain
google-auth
gemini-api
